DX11 전환을 위한 WinAPI 렌더링 제거

+++ BlueSource/BIndexBuffer.cpp
@@ -12,7 +12,7 @@ namespace blue::graphics
 
 	bool IndexBuffer::Create(const std::vector<UINT>& indices)
 	{
-		desc.ByteWidth = sizeof(UINT) * indices.size();
+		desc.ByteWidth = sizeof(UINT) * (UINT)indices.size();
 		desc.BindFlags = D3D11_BIND_FLAG::D3D11_BIND_INDEX_BUFFER;
 		desc.Usage = D3D11_USAGE_DEFAULT;
 		desc.CPUAccessFlags = 0;
