+++ BlueSource/BInput.cpp
@@ -103,11 +103,11 @@ namespace blue
 		mMousePosition.x = -1.0f;
 		mMousePosition.y = -1.0f;
 
-		if (mousePos.x > 0 && mousePos.x < width)
-			mMousePosition.x = mousePos.x;
+		if ((UINT)mousePos.x > 0 && (UINT)mousePos.x < width)
+			mMousePosition.x = (float)mousePos.x;
 
-		if (mousePos.y > 0 && mousePos.y < height)
-			mMousePosition.y = mousePos.y;
+		if ((UINT)mousePos.y > 0 && (UINT)mousePos.y < height)
+			mMousePosition.y = (float)mousePos.y;
 
 	}
