+++ BlueSource/BVertexBuffer.cpp
@@ -12,7 +12,7 @@ namespace blue::graphics
 
 	bool VertexBuffer::Create(const std::vector<Vertex>& vertexes)
 	{
-		desc.ByteWidth = sizeof(Vertex) * vertexes.size();
+		desc.ByteWidth = sizeof(Vertex) * (UINT)vertexes.size();
 		desc.BindFlags = D3D11_BIND_FLAG::D3D11_BIND_VERTEX_BUFFER;
 		desc.Usage = D3D11_USAGE::D3D11_USAGE_DYNAMIC;
 		desc.CPUAccessFlags = D3D11_CPU_ACCESS_FLAG::D3D11_CPU_ACCESS_WRITE;
